requirements.txt
fastapi~=0.63.0
ffmpeg-python
setuptools~=56.0.0